Trigger workflow for releases
Workflow triggers for pull request but the github.refs for pull request does not match contain string refs/heads/release- as value for the github.refs for pull request it refs/pull/pr-number/merge. Updated workflow to trigger on push on a branch release-*. This change is specific to release of 0.8.4. Implementation is not good to merge the code